# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Vroom Forecast API",
        "description": "Predicts vehicle reservations \u2014 powered by Ray Serve.",
        "version": "0.2.0"
    },
    "paths": {
...
        "/vehicles": {
            "get": {
                "summary": "List Vehicles Endpoint",
                "operationId": "list_vehicles_endpoint_vehicles_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/VehicleRecord"
                                    },
                                    "type": "array",
                                    "title": "Response List Vehicles Endpoint Vehicles Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Save Vehicle Endpoint",
                "operationId": "save_vehicle_endpoint_vehicles_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VehicleFeatures"
                            },
                            "example": {}
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SaveVehicleResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/vehicles/{vehicle_id}/features": {
            "get": {
                "summary": "Get Vehicle Features",
                "description": "Get the computed features from the online store for a saved vehicle.",
                "operationId": "get_vehicle_features_vehicles__vehicle_id__features_get",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Vehicle Id"
                        },
                        "example": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ComputedFeatures"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/predict": {
            "post": {
                "summary": "Predict Single",
                "operationId": "predict_single_predict_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VehicleFeatures"
                            },
                            "example": {}
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PredictionResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/predict/id": {
            "post": {
                "summary": "Predict By Id",
                "operationId": "predict_by_id_predict_id_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VehicleIdRequest"
                            },
                            "example": {}
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PredictionResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/predict/batch": {
            "post": {
                "summary": "Predict Batch",
                "operationId": "predict_batch_predict_batch_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "items": {
                                    "$ref": "#/components/schemas/VehicleFeatures"
                                },
                                "type": "array",
                                "title": "Vehicles"
                            },
                            "example": {}
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BatchPredictionResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/benchmark": {
            "post": {
                "summary": "Benchmark",
                "description": "Benchmark: feature computation (on the fly) + inference.",
                "operationId": "benchmark_benchmark_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BenchmarkRequest"
                            },
                            "example": {}
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BenchmarkResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/benchmark/id": {
            "post": {
                "summary": "Benchmark By Id",
                "description": "Benchmark: feature lookup (online store) + inference.",
                "operationId": "benchmark_by_id_benchmark_id_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BenchmarkByIdRequest"
                            },
                            "example": {}
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BenchmarkResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "SaveVehicleResponse": {
                "properties": {
                    "vehicle_id": {
                        "type": "integer",
                        "title": "Vehicle Id"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "event_published": {
                        "type": "boolean",
                        "title": "Event Published",
                        "default": true
                    }
                },
                "type": "object",
                "required": [
                    "vehicle_id",
                    "status"
                ],
                "title": "SaveVehicleResponse"
            },
...
            "VehicleRecord": {
                "properties": {
                    "vehicle_id": {
                        "type": "integer",
                        "title": "Vehicle Id"
                    },
                    "technology": {
                        "type": "integer",
                        "title": "Technology"
                    },
                    "actual_price": {
                        "type": "number",
                        "title": "Actual Price"
                    },
                    "recommended_price": {
                        "type": "number",
                        "title": "Recommended Price"
                    },
                    "num_images": {
                        "type": "integer",
                        "title": "Num Images"
                    },
                    "street_parked": {
                        "type": "integer",
                        "title": "Street Parked"
                    },
                    "description": {
                        "type": "integer",
                        "title": "Description"
                    },
                    "source": {
                        "type": "string",
                        "title": "Source",
                        "default": "csv"
                    },
                    "num_reservations": {
                        "type": "integer",
                        "title": "Num Reservations",
                        "default": 0
                    }
                },
                "type": "object",
                "required": [
                    "vehicle_id",
                    "technology",
                    "actual_price",
                    "recommended_price",
                    "num_images",
                    "street_parked",
                    "description"
                ],
                "title": "VehicleRecord",
                "description": "A stored vehicle with its ID, attributes, source, and reservation count."
            }
        }
    }
}